requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.16.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
blosc2==2.7.1
brotlipy==0.7.0
cachetools==5.5.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1734380492396/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet @ file:///home/builder/ci_310/chardet_1640804867535/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
clarabel==0.9.0
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
Columnar==1.4.1
comm==0.2.2
conda==22.11.1
conda-build==3.23.3
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
contourpy==1.3.1
cryptography @ file:///croot/cryptography_1665612644927/work
cupy-cuda11x==13.3.0
cvxpy==1.6.0
cycler==0.12.1
databricks-cli==0.18.0
databricks-sdk==0.42.0
debugpy==1.8.12
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
dnspython==2.2.1
docker==6.1.3
docker-pycreds==0.4.0
docopt-ng==0.9.0
entrypoints==0.4
exceptiongroup==1.0.4
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
expecttest==0.1.4
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.17.0
fire==0.7.0
Flask==2.2.5
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.55.8
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
future==0.18.2
gitdb==4.0.12
GitPython==3.1.44
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==20.1.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperopt==0.1.2
hypothesis==6.61.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==5.2.0
ipykernel==6.29.5
ipython @ file:///croot/ipython_1670919316550/work
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.0.3
joblib==1.4.2
json5==0.10.0
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
lightgbm @ file:///croot/lightgbm_1723565407197/work
lightning-utilities==0.11.9
loguru==0.7.3
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune==3.1.1
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mlflow==1.30.0
mlflow-skinny==2.20.1
mpmath==1.2.1
msgpack==1.1.0
multidict==6.1.0
munch==4.0.0
narwhals==1.24.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.9.2
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.23.5
oauthlib==3.2.2
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
osqp==0.6.7.post3
overrides==7.7.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
patsy==1.0.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
pkginfo @ file:///croot/pkginfo_1666725041340/work
platformdirs==4.3.6
plotly==6.0.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.25.6
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.6
pydantic_core==2.27.2
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT==2.10.1
pymongo==3.7.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.2.1
pyqlib @ file:///workspace/MATCC/qlib
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-etcd==0.4.5
python-json-logger==3.2.1
python-redis-lock==4.0.0
python-socketio==5.12.1
pytorch-lightning==1.9.5
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML==6.0
pyzmq==26.2.1
qdldl==0.1.7.post5
querystring-parser==1.2.4
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.17.36
ruamel.yaml.clib==0.2.12
sacred==0.8.7
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.1
scs==3.2.7.post2
Send2Trash==1.8.3
sentry-sdk==2.20.0
setproctitle==1.3.4
simple-websocket==1.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve @ file:///croot/soupsieve_1666296392845/work
SQLAlchemy==1.4.54
sqlparse==0.5.3
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
statsmodels==0.14.4
string-color==1.2.3
sympy==1.11.1
tables==3.10.1
tabulate==0.9.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.2.1
toolz @ file:///croot/toolz_1667464077321/work
torch==1.13.1
torchelastic==0.2.2
torchmetrics==1.5.2
torchpq==0.3.0.1
torchtext==0.14.1
torchvision==0.14.1
tornado==6.4.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets @ file:///croot/traitlets_1671143879854/work
types-dataclasses==0.6.6
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1670526988650/work
-e git+https://github.com/minyoungg/vqtorch@02e60a19bd742c17b0bf3e1925f23796d54cbeac#egg=vqtorch
wandb==0.19.5
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
yarl==1.18.3
zipp==3.21.0
